+++ src/Repository/Model/Artist.php
@@ -924,8 +924,8 @@ class Artist extends database_object implements library_item, GarbageCollectible
      */
     public static function update_artist_map($artist_id, $object_type, $object_id)
     {
-        if ((int)$artist_id > 0) {
-            debug_event(__CLASS__, "$object_type update_artist_map $artist_id: $object_id", 5);
+        if ((int)$artist_id > 0 && (int)$object_id > 0) {
+            debug_event(__CLASS__, "update_artist_map artist_id {" . $artist_id . "} $object_type {" . $object_id . "}", 5);
             $sql = "INSERT IGNORE INTO `artist_map` (`artist_id`, `object_type`, `object_id`) VALUES (?, ?, ?);";
             Dba::write($sql, array($artist_id, $object_type, $object_id));
         }
@@ -936,8 +936,8 @@ class Artist extends database_object implements library_item, GarbageCollectible
      */
     public static function remove_artist_map($artist_id, $object_type, $object_id)
     {
-        if ((int)$artist_id > 0) {
-            debug_event(__CLASS__, "$object_type remove_artist_map $artist_id: $object_id", 5);
+        if ((int)$artist_id > 0 && (int)$object_id > 0) {
+            debug_event(__CLASS__, "remove_artist_map artist_id {" . $artist_id . "} $object_type {" . $object_id . "}", 5);
             $sql = "DELETE FROM `artist_map` WHERE `artist_id` = ? AND `object_type` = ? AND `object_id` = ?;";
             Dba::write($sql, array($artist_id, $object_type, $object_id));
         }
